+++ src/ServiceProvider.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace StatamicRadPack\meilisearch;
+namespace StatamicRadPack\Meilisearch;
 
 use Illuminate\Foundation\Application;
-use meilisearch\Client;
+use Meilisearch\Client;
 use Statamic\Facades\Search;
 use Statamic\Providers\AddonServiceProvider;
 
@@ -11,16 +11,17 @@ class ServiceProvider extends AddonServiceProvider
 {
     public function bootAddon()
     {
-        Search::extend('meilisearch', function (Application $app, array $config, $name) {
+        Search::extend('meilisearch', function (Application $app, array $config, $name, $locale = null) {
             $client = $app->makeWith(Client::class, [
                 'url' => $config['credentials']['url'],
                 'apiKey' => $config['credentials']['secret'],
             ]);
 
-            return $app->makeWith(meilisearch\Index::class, [
+            return $app->makeWith(Meilisearch\Index::class, [
                 'client' => $client,
                 'name' => $name,
                 'config' => $config,
+                'locale' => $locale,
             ]);
         });
     }